Add --keep argument to test_shaders.py.
Improves debuggability when regressions are found.

+++ test_shaders.py
@@ -86,21 +86,22 @@ def make_reference_dir(path):
     if not os.path.exists(base):
         os.makedirs(base)
 
-def regression_check(shader, glsl, update):
+def regression_check(shader, glsl, update, keep):
     reference = os.path.join('./reference', shader)
     if os.path.exists(reference):
         if md5_for_file(glsl) != md5_for_file(reference):
-            print('Generated GLSL has changed for {}!'.format(reference))
-
             if update:
+                print('Generated GLSL has changed for {}!'.format(reference))
                 # If we expect changes, update the reference file.
                 if os.path.exists(reference):
                     os.remove(reference)
                 make_reference_dir(reference)
                 shutil.move(glsl, reference)
             else:
-                # Otherwise, fail the test.
-                os.remove(glsl)
+                print('Generated GLSL in {} does not match reference {}!'.format(glsl, reference))
+                # Otherwise, fail the test. Keep the shader file around so we can inspect.
+                if not keep:
+                    os.remove(glsl)
                 sys.exit(1)
         else:
             os.remove(glsl)
@@ -109,14 +110,14 @@ def regression_check(shader, glsl, update):
         make_reference_dir(reference)
         shutil.move(glsl, reference)
 
-def test_shader(stats, shader, update):
+def test_shader(stats, shader, update, keep):
     print('Testing shader:', shader)
     spirv, glsl = cross_compile(shader)
 
     if stats:
         cross_stats = get_shader_stats(glsl)
 
-    regression_check(shader, glsl, update)
+    regression_check(shader, glsl, update, keep)
     os.remove(spirv)
 
     if stats:
@@ -130,19 +131,19 @@ def test_shader(stats, shader, update):
             a.append(str(i))
         print(','.join(a), file = stats)
 
-def test_shaders(shader_dir, update, malisc):
+def test_shaders(shader_dir, update, malisc, keep):
     if malisc:
         with open('stats.csv', 'w') as stats:
             print('Shader,OrigRegs,OrigUniRegs,OrigALUShort,OrigLSShort,OrigTEXShort,OrigALULong,OrigLSLong,OrigTEXLong,CrossRegs,CrossUniRegs,CrossALUShort,CrossLSShort,CrossTEXShort,CrossALULong,CrossLSLong,CrossTEXLong', file = stats)
             for f in os.walk(os.path.join(shader_dir)):
                 for i in f[2]:
                     shader = os.path.join(f[0], i)
-                    test_shader(stats, shader, update)
+                    test_shader(stats, shader, update, keep)
     else:
         for f in os.walk(os.path.join(shader_dir)):
             for i in f[2]:
                 shader = os.path.join(f[0], i)
-                test_shader(None, shader, update)
+                test_shader(None, shader, update, keep)
 
 def main():
     parser = argparse.ArgumentParser(description = 'Script for regression testing.')
@@ -151,6 +152,9 @@ def main():
     parser.add_argument('--update',
             action = 'store_true',
             help = 'Updates reference files if there is a mismatch. Use when legitimate changes in output is found.')
+    parser.add_argument('--keep',
+            action = 'store_true',
+            help = 'Leave failed GLSL shaders on disk if they fail regression. Useful for debugging.')
     parser.add_argument('--malisc',
             action = 'store_true',
             help = 'Use malisc offline compiler to determine static cycle counts before and after spir2cross.')
@@ -160,9 +164,10 @@ def main():
         sys.stderr.write('Need shader folder.\n')
         sys.exit(1)
 
-    test_shaders(args.folder, args.update, args.malisc)
+    test_shaders(args.folder, args.update, args.malisc, args.keep)
     if args.malisc:
         print('Stats in stats.csv!')
+    print('Tests completed!')
 
 if __name__ == '__main__':
     main()
